Revert back to pycrypto instead of cryptography
Cryptography seems to be the better library overall, but I'm tired of
installation headaches about finding OpenSSL headers, etc. The main
reason I switched to crytography in the first place was supporting pypy,
but as of pycookiecheat v0.4.0 I will no longer support python2, so pypy
will also not work.