Add Firefox support via firefox_cookies()
Add a new `firefox_cookies()`, similar to `chrome_cookies()`. The
arguments are slightly different:

  - `profile_name` instead of `cookie_file`, since Firefox supports
    multiple profiles in the default config directory. `profile_name`
    may be a glob pattern because Firefox profiles have a random
    prefix that's not directly user-visible or -maintained.
  - No `password` argument, since Firefox doesn't encrypt its cookies.

Firefox stores its cookies in an SQLite3 database file. While Firefox
is running, the database is locked, so the code makes a temporary
copy of it before reading cookies from it.

If no `profile_name` is given the default profile will be returned, by
parsing the `profiles.ini` file in the Firefox config directory. This
can be overridden by passing a profile name or pattern to
`profile_name`.

A note on the unit tests: They don't test the scenario where Firefox
is running while getting the cookies. This is because the synchronous
playwright API doesn't support yielding from inside the running `with`
context manager. The asynchronous API probably does, but I haven't
looked into using that. Nonetheless the code supports retrieving
cookies while Firefox is running.

Closes n8henrie/pycookiecheat#48.