Provide an enum to ensure browser variants
This helps provide a single entry and exit point for library users to
ensure that any strings passed in are parsed into exactly one browser
type.

Will implement for Firefox shortly.

This is partly in preparation for adding a CLI script, at which point
this should be the only part of the app that accepts a user-provided
string for a browser name; all other parts of the API should be passing
an instance of `BrowserType`.