Provide a pathway to return list[Cookie] instead of just dict
Returning a `dict` loses a lot of information from the `Cookie`.
However, the `dict` is what we've previously returned and works well
with e.g. Requests. Instead of breaking backward compatibility here,
I think it's reasonable to provide an option to return the entire
`list[Cookie]` with a new parameter. This should make it easier for
external tooling if they need the full content of the cookie, but
preserve backward compatibility, at the cost of an unusal return type
and a few typing changes.