Add new top-level `get_cookies`
`get_cookies` calls out to `chrome_cookies` or `firefox_cookies` based
on `browser`.

Because `browser` needs to be specified anyway, this should make it
easier for users to have a single entrypoint.

As part of this refactor I changed argument order to be more similar
between the two browser-specific functions and required everything other
than `url` to be keyword-only, to reduce ambiguity, which required some
changes in the Firefox tests.